Change UV index to Double

+++ app/src/main/java/com/davidtakac/bura/forecast/cache/ConvertCacheJsonToForecast.kt
@@ -181,7 +181,7 @@ private fun jsonToUvIndexMoment(jsonObject: JSONObject): UvIndexMoment =
     convertJsonToMoment(jsonObject) {
         UvIndexMoment(
             hour = it,
-            uvIndex = UvIndex(value = jsonObject.getInt(CacheJsonSerialNames.UV_INDEX_VALUE))
+            uvIndex = UvIndex(value = jsonObject.getDouble(CacheJsonSerialNames.UV_INDEX_VALUE))
         )
     }
